refactor(http-py): split config support modules
